Add support for custom segment colors
Sometimes you need to change the segments colors and do not want to rely
on the `d3InterpolateHSL()` function. You should be able to provide the
colors as a prop.

If the user now provides the `segmentColors` prop as an array of colors,
and it has the same length as the number of segments, we use the
provided colors for the segments.

It could be possible to change the code to provide another color
function, but since this wasn't necessary this time, I didn't add it.